fix: long cast

+++ src/main/java/org/fatec/esportiva/request/ExchangeVoucherDto.java
@@ -21,7 +21,7 @@ public class ExchangeVoucherDto {
     private Long id;
 
     @NotNull(message = "Valor do cupom de desconto não pode ficar em branco")
-    @Min(value = 0.01, message = "Valor do cupom de desconto não pode ser negativo ou zero")
+    @Min(value = (long) 0.01, message = "Valor do cupom de desconto não pode ser negativo ou zero")
     private BigDecimal value;
 
     @NotNull(message = "Quantidade de cupons ficar em branco")
